Moved do_stream_reply from client to intercom
